Move GetCapabilities extended parameters to the user code
