+++ ogc-example/src/main/scala/geotrellis/server/ogc/wcs/WcsView.scala
@@ -17,6 +17,7 @@
 package geotrellis.server.ogc.wcs
 
 import geotrellis.server.ogc.params.ParamError
+import geotrellis.server.ogc.ows.OwsDataRecord
 
 import org.backuity.ansi.AnsiFormatter.FormattedHelper
 import org.http4s.scalaxml._
@@ -24,11 +25,48 @@ import org.http4s._
 import org.http4s.dsl.io._
 import cats.effect._
 import cats.data.Validated
+import cats.syntax.option._
+import opengis.ows.{AllowedValues, AnyValue, DomainType, ValueType}
+import org.log4s.getLogger
+import opengis._
+import scalaxb._
 
 import java.net._
 
 class WcsView(wcsModel: WcsModel, serviceUrl: URL) {
-  val logger = org.log4s.getLogger
+  val logger = getLogger
+
+  val extendedParameters: List[DomainType] = DomainType(
+    possibleValuesOption1 = OwsDataRecord(
+      AllowedValues(
+        OwsDataRecord(
+          ValueType("all")
+        ) :: OwsDataRecord(
+          ValueType("data")
+        ) :: OwsDataRecord(
+          ValueType("nodata")
+        ) :: Nil)
+    ),
+    DefaultValue = ValueType("all").some,
+    attributes = Map(
+      "@name" -> DataRecord("target")
+    )
+  ) :: DomainType(
+    possibleValuesOption1 = OwsDataRecord(AnyValue()),
+    attributes = Map(
+      "@name" -> DataRecord("zFactor")
+    )
+  ) :: DomainType(
+    possibleValuesOption1 = OwsDataRecord(AnyValue()),
+    attributes = Map(
+      "@name" -> DataRecord("azimuth")
+    )
+  ) :: DomainType(
+    possibleValuesOption1 = OwsDataRecord(AnyValue()),
+    attributes = Map(
+      "@name" -> DataRecord("altitude")
+    )
+  ) :: Nil
 
   private def handleError[Result: EntityEncoder[IO, *]](result: Either[Throwable, Result]): IO[Response[IO]] = result match {
     case Right(res) =>
@@ -52,7 +90,7 @@ class WcsView(wcsModel: WcsModel, serviceUrl: URL) {
         wcsParams match {
           case p: GetCapabilitiesWcsParams =>
             logger.debug(ansi"%bold{GetCapabilities: $serviceUrl}")
-            Ok(new CapabilitiesView(wcsModel, serviceUrl).toXML)
+            Ok(new CapabilitiesView(wcsModel, serviceUrl, extendedParameters).toXML)
 
           case p: DescribeCoverageWcsParams =>
             logger.debug(ansi"%bold{DescribeCoverage: ${req.uri}}")
