+++ ogc-example/src/main/scala/geotrellis/server/ogc/wms/WmsView.scala
@@ -18,6 +18,7 @@ package geotrellis.server.ogc.wms
 
 import geotrellis.server._
 import geotrellis.server.ogc._
+import geotrellis.server.ogc.utils._
 import geotrellis.server.ogc.params.ParamError
 import geotrellis.server.ogc.wms.WmsParams.{GetCapabilities, GetMap}
 
@@ -25,24 +26,52 @@ import geotrellis.raster.RasterExtent
 import geotrellis.raster._
 import com.azavea.maml.error._
 import com.azavea.maml.eval._
-import scalaxb.CanWriteXML
 import org.http4s.scalaxml._
 import org.http4s.circe._
 import org.http4s._
 import org.http4s.dsl.io._
-import org.http4s.implicits._
 import _root_.io.circe.syntax._
-import cats._
 import cats.implicits._
 import cats.effect._
 import cats.data.Validated._
 import com.github.blemale.scaffeine.{Cache, Scaffeine}
+import org.log4s.getLogger
+import opengis._
+import scalaxb._
 
 import java.net.URL
+
 import scala.concurrent.duration._
+import scala.xml.Elem
 
 class WmsView(wmsModel: WmsModel, serviceUrl: URL) {
-  val logger = org.log4s.getLogger
+  val logger = getLogger
+
+  val extendedCapabilities: List[DataRecord[Elem]] = {
+    val targetCells = DataRecord(
+      None, "target".some,
+      DataRecord("all").toXML ++
+      DataRecord("data").toXML ++
+      DataRecord("nodata").toXML ++
+      DataRecord(None, "default".some, "all").toXML
+    )
+
+    val focalHillshade: Elem = ExtendedElement(
+      "FocalHillshade",
+      DataRecord("zFactor"),
+      DataRecord("azimuth"),
+      DataRecord("altitude"),
+      targetCells
+    )
+
+    val focalSlope: Elem = ExtendedElement(
+      "FocalSlope",
+      DataRecord("zFactor"),
+      targetCells
+    )
+
+    ExtendedCapabilities(focalHillshade, focalSlope)
+  }
 
   private val histoCache: Cache[OgcLayer, Interpreted[List[Histogram[Double]]]] =
     Scaffeine()
@@ -59,7 +88,7 @@ class WmsView(wmsModel: WmsModel, serviceUrl: URL) {
         BadRequest(msg)
 
       case Valid(wmsReq: GetCapabilities) =>
-        Ok(new CapabilitiesView(wmsModel, serviceUrl).toXML)
+        Ok(new CapabilitiesView(wmsModel, serviceUrl, extendedCapabilities).toXML)
 
       case Valid(wmsReq: GetMap) =>
         val re = RasterExtent(wmsReq.boundingBox, wmsReq.width, wmsReq.height)
