+++ ogc/src/main/scala/geotrellis/server/ogc/wcs/CapabilitiesView.scala
@@ -32,7 +32,8 @@ import java.net.{URI, URL}
 
 class CapabilitiesView(
   wcsModel: WcsModel,
-  serviceUrl: URL
+  serviceUrl: URL,
+  extendedParameters: List[DomainType] = Nil
 ) {
   def toXML: Elem = {
     val serviceIdentification = ServiceIdentification(
@@ -132,7 +133,7 @@ class CapabilitiesView(
               )) :: Nil)
           )
         ) :: Nil,
-        Parameter = DomainType(
+        Parameter = (DomainType(
           possibleValuesOption1 = OwsDataRecord(
             AllowedValues(
               OwsDataRecord(
@@ -161,37 +162,7 @@ class CapabilitiesView(
           ),
           DefaultValue = ValueType("nearest neighbor").some,
           attributes = Map("@name" -> DataRecord("InterpolationType"))
-        ) :: DomainType(
-          possibleValuesOption1 = OwsDataRecord(
-            AllowedValues(
-              OwsDataRecord(
-                ValueType("all")
-              ) :: OwsDataRecord(
-                ValueType("data")
-              ) :: OwsDataRecord(
-                ValueType("nodata")
-              ) :: Nil)
-          ),
-          DefaultValue = ValueType("all").some,
-          attributes = Map(
-            "@name" -> DataRecord("target")
-          )
-        ) :: DomainType(
-          possibleValuesOption1 = OwsDataRecord(AnyValue()),
-          attributes = Map(
-            "@name" -> DataRecord("zFactor")
-          )
-        ) :: DomainType(
-          possibleValuesOption1 = OwsDataRecord(AnyValue()),
-          attributes = Map(
-            "@name" -> DataRecord("azimuth")
-          )
-        ) :: DomainType(
-          possibleValuesOption1 = OwsDataRecord(AnyValue()),
-          attributes = Map(
-            "@name" -> DataRecord("altitude")
-          )
-        ) :: Nil,
+        ) :: Nil) ::: extendedParameters,
         attributes = Map("@name" -> DataRecord("GetCoverage"))
       )
